chore(xo-server): move RPU logic from API layer to XenServers mixin
