+++ packages/xo-server/src/xo-mixins/xen-servers.mjs
@@ -4,6 +4,10 @@ import pick from 'lodash/pick.js'
 import { asyncEach } from '@vates/async-each'
 import { BaseError } from 'make-error'
 import { createLogger } from '@xen-orchestra/log'
+import { createPredicate } from 'value-matcher'
+import { decorateClass } from '@vates/decorate-with'
+import { defer } from 'golike-defer'
+import { extractIdsFromSimplePattern } from '@xen-orchestra/backups/extractIdsFromSimplePattern.js'
 import { fibonacci } from 'iterable-backoff'
 import { networkInterfaces } from 'os'
 import { noSuchObject } from 'xo-common/api-errors.js'
@@ -39,7 +43,7 @@ const log = createLogger('xo:xo-mixins:xen-servers')
 // Server is connected:
 // - _xapis[server.id] id defined
 // - _serverIdsByPool[xapi.pool.$id] is server.id
-export default class {
+export default class XenServers {
   constructor(app, { safeMode }) {
     this._objectConflicts = { __proto__: null } // TODO: clean when a server is disconnected.
     const serversDb = (this._servers = new Servers({
@@ -632,4 +636,58 @@ export default class {
       })
       ::ignoreErrors()
   }
+
+  async rollingPoolUpdate($defer, pool) {
+    const app = this._app
+
+    const [schedules, jobs] = await Promise.all([app.getAllSchedules(), app.getAllJobs('backup')])
+
+    const poolId = pool.id
+
+    const jobsOfthePool = []
+    jobs.forEach(({ id: jobId, vms }) => {
+      if (vms.id !== undefined) {
+        for (const vmId of extractIdsFromSimplePattern(vms)) {
+          // try/catch to avoid `no such object`
+          try {
+            if (app.getObject(vmId).$poolId === poolId) {
+              jobsOfthePool.push(jobId)
+              break
+            }
+          } catch {}
+        }
+      } else {
+        // Smart mode
+        // For smart mode, we take a simplified approach:
+        // - if smart mode is explicitly 'resident' or 'not resident' on pools, we
+        //   check if it concerns this pool
+        // - if not, the job may concern this pool so we add it to `jobsOfThePool`
+        if (vms.$pool === undefined || createPredicate(vms.$pool)(poolId)) {
+          jobsOfthePool.push(jobId)
+        }
+      }
+    })
+
+    // Disable schedules
+    await Promise.all(
+      schedules
+        .filter(schedule => jobsOfthePool.includes(schedule.jobId) && schedule.enabled)
+        .map(async schedule => {
+          await app.updateSchedule({ ...schedule, enabled: false })
+          $defer(() => app.updateSchedule({ ...schedule, enabled: true }))
+        })
+    )
+
+    // Disable load balancer
+    if ((await app.getOptionalPlugin('load-balancer'))?.loaded) {
+      await app.unloadPlugin('load-balancer')
+      $defer(() => app.loadPlugin('load-balancer'))
+    }
+
+    await this.getXapi(pool).rollingPoolUpdate()
+  }
 }
+
+decorateClass(XenServers, {
+  rollingPoolUpdate: defer,
+})
